requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
flask==2.3.3
SQLAlchemy==2.0.23
mysqlclient==2.2.0
